# swagger/swagger.yaml
swagger: "2.0"
info:
  version: '1.0.0'
  title: Swagger sam-api
...
  contact:
    email: norbert.bondarczuk@wipro.com
host: localhost:8000
basePath: /api
securityDefinitions: {}
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /order:
    post:
      description: >-
        A new mapping is created in working status, unconfirmed. The entries segment value must be validated with values from dictionary Segment.

        Requires:

        - Control role.
      summary: OrderCreateOne
      tags:
      - order
      operationId: OrderPost
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: body
        in: body
        required: true
        description: Creation of object Order
        schema:
          $ref: '#/definitions/RequestSetOrder'
      responses:
        201:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOrder'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/ResultSetError'						
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
    get:
      description: Set of orders is read and returned. The records are active ones, ie. status W or C for release 0 or status P for max release_id.
      summary: OrderReadActiveAll
      tags:
      - order
      operationId: OrderReadActiveAll
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: Content-Type
        in: header
        required: true
        type: string
        enum:
        - application/json
        - application/xmlx
        - application/csv		
        description: 'Determines response payload format'		
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOrders'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'   
    delete:
      description: Purges whole resource.
      summary: OrderDeleteAll
      tags:
      - order
      operationId: OrderDeleteAll
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOrders'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'              
...
  /order/log:
    get:
      description: A set of accout order logs is returned
      summary: OrderReadLog
      tags:
      - order
      operationId: OrderReadLog
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: account
        in: path
        required: true
        type: string
        description: BSCS account id for which log is produced by triggers
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetOrderLogs'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'			
...
definitions:
...
  RequestSetOrder:
    title: RequestSetOrder
    type: object
    properties:
      data:
        $ref: '#/definitions/Order'
  ResultSetOrder:
    title: ResultSetOrder
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      data:
        $ref: '#/definitions/Order'
  Order:
    title: Order
    example:
      status: W
      releaseId: 0
      bscsAccount: BSCSACCOUNT
      segmentCode: XXX
      orderNumber: 123
      validFromDate: 2017-07-01
    type: object
    properties:
      status:
        type: string
      releaseId:
        type: integer
        format: int32
      bscsAccount:
        type: string
      segmentCode:
        type: string
      orderNumber:
        type: string
      validFromDate:
        type: string
        format: date
      entryDate:
        type: string
        format: date-time
      entryOwner:
        type: string
      updateDate:
        type: string
        format: date-time
      updateOwner:
        type: string
      releaseDate:
        type: string
        format: date-time
      releaseOwner:
        type: string
  OrderLog:
    title: OrderLog
    example:
      opcode: I
      opdate: 2017-07-21T17:32:28Z
      status: W
      releaseId: 0
      bscsAccount: BSCSACCOUNT
      segmentCode: XXX
      orderNumber: 123
      validFromDate: 2017-07-01
    type: object
    properties:
      opcode:
        type: string
      opdate:
        type: string
        format: int32			
      status:
        type: string
      releaseId:
        type: integer
        format: int32
      bscsAccount:
        type: string
      segmentCode:
        type: string
      orderNumber:
        type: string
      validFromDate:
        type: string
        format: date
      entryDate:
        type: string
        format: date-time
      entryOwner:
        type: string
      updateDate:
        type: string
        format: date-time
      updateOwner:
        type: string
      releaseDate:
        type: string
        format: date-time
      releaseOwner:
        type: string		
  ResultSetOrders:
    title: ResultSetOrders
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      count:
        type: integer
        format: int32
      data:
        type: array
        items:
          $ref: '#/definitions/Order'
  ResultSetOrderLogs:
    title: ResultSetOrderLogs
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      count:
        type: integer
        format: int32
      data:
        type: array
        items:
          $ref: '#/definitions/OrderLog'		  
...
  ResultSetError:
    title: ResultSetError
    type: object
    properties:
      status:
        $ref: '#/definitions/Status12'
      message:
        type: string
...
  Status:
    title: Status
    example: Ok
    type: string
    enum:
    - Ok
  Status12:
    title: Status12
    example: Error
    type: string
    enum:
    - Error
...